Remove CodeGen_OpenGL_Dev
5626 removed OpenGL support, but didn't remove this no-longer-needed class; removed it here. Moved CodeGen_GLSLBase into CodGen_OpenGLCompute_Dev.cpp (which is now the only subclass), but didn't yet attempt to consolidate them into a single class (not sure if any of that logic might be used/useful in the not-yet-done Vulkan backend).
