+++ src/CMakeLists.txt
@@ -33,7 +33,6 @@ set(HEADER_FILES
     CodeGen_Metal_Dev.h
     CodeGen_MIPS.h
     CodeGen_OpenCL_Dev.h
-    CodeGen_OpenGL_Dev.h
     CodeGen_OpenGLCompute_Dev.h
     CodeGen_Posix.h
     CodeGen_PowerPC.h
@@ -197,7 +196,6 @@ set(SOURCE_FILES
     CodeGen_Metal_Dev.cpp
     CodeGen_MIPS.cpp
     CodeGen_OpenCL_Dev.cpp
-    CodeGen_OpenGL_Dev.cpp
     CodeGen_OpenGLCompute_Dev.cpp
     CodeGen_Posix.cpp
     CodeGen_PowerPC.cpp
